refactor(python): split `read_database` functionality into cleaner module structure
