+++ py-polars/polars/io/database/__init__.py
@@ -1,0 +1,6 @@
+from polars.io.database.functions import read_database, read_database_uri
+
+__all__ = [
+    "read_database",
+    "read_database_uri",
+]
