+++ py-polars/polars/io/database/_executor.py
@@ -1,0 +1,498 @@
+from __future__ import annotations
+
+import re
+import warnings
+from collections.abc import Coroutine
+from contextlib import suppress
+from inspect import Parameter, isclass, signature
+from typing import TYPE_CHECKING, Any, Iterable, Sequence
+
+from polars.convert import from_arrow
+from polars.datatypes import (
+    INTEGER_DTYPES,
+    N_INFER_DEFAULT,
+    UNSIGNED_INTEGER_DTYPES,
+    Decimal,
+    Float32,
+    Float64,
+)
+from polars.datatypes.convert import _map_py_type_to_dtype
+from polars.exceptions import UnsuitableSQLError
+from polars.io.database._arrow_registry import ARROW_DRIVER_REGISTRY
+from polars.io.database._cursor_proxies import ODBCCursorProxy
+from polars.io.database._inference import (
+    _infer_dtype_from_database_typename,
+    _integer_dtype_from_nbits,
+)
+
+if TYPE_CHECKING:
+    import sys
+    from types import TracebackType
+
+    import pyarrow as pa
+
+    from polars.io.database._arrow_registry import ArrowDriverProperties
+
+    if sys.version_info >= (3, 10):
+        from typing import TypeAlias
+    else:
+        from typing_extensions import TypeAlias
+
+    if sys.version_info >= (3, 11):
+        from typing import Self
+    else:
+        from typing_extensions import Self
+
+    from polars import DataFrame
+    from polars.datatypes import PolarsDataType
+    from polars.type_aliases import ConnectionOrCursor, Cursor, SchemaDict
+
+    try:
+        from sqlalchemy.sql.expression import Selectable
+    except ImportError:
+        Selectable: TypeAlias = Any  # type: ignore[no-redef]
+
+    from sqlalchemy.sql.elements import TextClause
+
+
+_INVALID_QUERY_TYPES = {
+    "ALTER",
+    "ANALYZE",
+    "CREATE",
+    "DELETE",
+    "DROP",
+    "INSERT",
+    "REPLACE",
+    "UPDATE",
+    "UPSERT",
+    "USE",
+    "VACUUM",
+}
+
+
+class ConnectionExecutor:
+    """Abstraction for querying databases with user-supplied connection objects."""
+
+    # indicate if we can/should close the cursor on scope exit. note that we
+    # should never close the underlying connection, or a user-supplied cursor.
+    can_close_cursor: bool = False
+
+    def __init__(self, connection: ConnectionOrCursor) -> None:
+        self.driver_name = (
+            "arrow_odbc_proxy"
+            if isinstance(connection, ODBCCursorProxy)
+            else type(connection).__module__.split(".", 1)[0].lower()
+        )
+        self.cursor = self._normalise_cursor(connection)
+        self.result: Any = None
+
+    def __enter__(self) -> Self:
+        return self
+
+    def __exit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc_val: BaseException | None,
+        exc_tb: TracebackType | None,
+    ) -> None:
+        # if we created it and are finished with it, we can
+        # close the cursor (but NOT the connection)
+        if type(self.cursor).__name__ == "AsyncConnection":
+            self._run_async(self._close_async_cursor())
+        elif self.can_close_cursor and hasattr(self.cursor, "close"):
+            self.cursor.close()
+
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} module={self.driver_name!r}>"
+
+    async def _close_async_cursor(self) -> None:
+        if self.can_close_cursor and hasattr(self.cursor, "close"):
+            from sqlalchemy.ext.asyncio.exc import AsyncContextNotStarted
+
+            with suppress(AsyncContextNotStarted):
+                await self.cursor.close()
+
+    def _fetch_arrow(
+        self,
+        driver_properties: ArrowDriverProperties,
+        *,
+        batch_size: int | None,
+        iter_batches: bool,
+    ) -> Iterable[pa.RecordBatch]:
+        """Yield Arrow data as a generator of one or more RecordBatches or Tables."""
+        fetch_batches = driver_properties["fetch_batches"]
+        if not iter_batches or fetch_batches is None:
+            fetch_method, sz = driver_properties["fetch_all"], []
+            if isinstance(fetch_method, tuple):
+                fetch_method, chunk_size = fetch_method
+                sz = [chunk_size]
+            yield getattr(self.result, fetch_method)(*sz)
+        else:
+            size = batch_size if driver_properties["exact_batch_size"] else None
+            repeat_batch_calls = driver_properties["repeat_batch_calls"]
+            fetchmany_arrow = getattr(self.result, fetch_batches)
+            if not repeat_batch_calls:
+                yield from fetchmany_arrow(size)
+            else:
+                while True:
+                    arrow = fetchmany_arrow(size)
+                    if not arrow:
+                        break
+                    yield arrow
+
+    @staticmethod
+    def _fetchall_rows(result: Cursor) -> Iterable[Sequence[Any]]:
+        """Fetch row data in a single call, returning the complete result set."""
+        rows = result.fetchall()
+        return (
+            [tuple(row) for row in rows]
+            if rows and not isinstance(rows[0], (list, tuple))
+            else rows
+        )
+
+    def _fetchmany_rows(
+        self, result: Cursor, batch_size: int | None
+    ) -> Iterable[Sequence[Any]]:
+        """Fetch row data incrementally, yielding over the complete result set."""
+        while True:
+            rows = result.fetchmany(batch_size)
+            if not rows:
+                break
+            elif isinstance(rows[0], (list, tuple)):
+                yield rows
+            else:
+                yield [tuple(row) for row in rows]
+
+    def _from_arrow(
+        self,
+        *,
+        batch_size: int | None,
+        iter_batches: bool,
+        schema_overrides: SchemaDict | None,
+        infer_schema_length: int | None,
+    ) -> DataFrame | Iterable[DataFrame] | None:
+        """Return resultset data in Arrow format for frame init."""
+        try:
+            for driver, driver_properties in ARROW_DRIVER_REGISTRY.items():
+                if re.match(f"^{driver}$", self.driver_name):
+                    fetch_batches = driver_properties["fetch_batches"]
+                    self.can_close_cursor = fetch_batches is None or not iter_batches
+                    frames = (
+                        from_arrow(batch, schema_overrides=schema_overrides)
+                        for batch in self._fetch_arrow(
+                            driver_properties,
+                            iter_batches=iter_batches,
+                            batch_size=batch_size,
+                        )
+                    )
+                    return frames if iter_batches else next(frames)  # type: ignore[arg-type,return-value]
+        except Exception as err:
+            # eg: valid turbodbc/snowflake connection, but no arrow support
+            # compiled in to the underlying driver (or on this connection)
+            arrow_not_supported = (
+                "does not support Apache Arrow",
+                "Apache Arrow format is not supported",
+            )
+            if not any(e in str(err) for e in arrow_not_supported):
+                raise
+
+        return None
+
+    def _from_rows(
+        self,
+        *,
+        batch_size: int | None,
+        iter_batches: bool,
+        schema_overrides: SchemaDict | None,
+        infer_schema_length: int | None,
+    ) -> DataFrame | Iterable[DataFrame] | None:
+        """Return resultset data row-wise for frame init."""
+        from polars import DataFrame
+
+        if is_async := isinstance(original_result := self.result, Coroutine):
+            self.result = self._run_async(self.result)
+        try:
+            if hasattr(self.result, "fetchall"):
+                if self.driver_name == "sqlalchemy":
+                    if hasattr(self.result, "cursor"):
+                        cursor_desc = {
+                            d[0]: d[1:] for d in self.result.cursor.description
+                        }
+                    elif hasattr(self.result, "_metadata"):
+                        cursor_desc = {k: None for k in self.result._metadata.keys}
+                    else:
+                        msg = f"Unable to determine metadata from query result; {self.result!r}"
+                        raise ValueError(msg)
+                else:
+                    cursor_desc = {d[0]: d[1:] for d in self.result.description}
+
+                schema_overrides = self._inject_type_overrides(
+                    description=cursor_desc,
+                    schema_overrides=(schema_overrides or {}),
+                )
+                result_columns = list(cursor_desc)
+                frames = (
+                    DataFrame(
+                        data=rows,
+                        schema=result_columns,
+                        schema_overrides=schema_overrides,
+                        infer_schema_length=infer_schema_length,
+                        orient="row",
+                    )
+                    for rows in (
+                        list(self._fetchmany_rows(self.result, batch_size))
+                        if iter_batches
+                        else [self._fetchall_rows(self.result)]  # type: ignore[list-item]
+                    )
+                )
+                return frames if iter_batches else next(frames)  # type: ignore[arg-type]
+            return None
+        finally:
+            if is_async:
+                original_result.close()
+
+    @staticmethod
+    def _run_async(co: Coroutine) -> Any:  # type: ignore[type-arg]
+        """Consolidate async event loop acquisition and coroutine/func execution."""
+        import asyncio
+
+        try:
+            with warnings.catch_warnings():
+                warnings.simplefilter("ignore", DeprecationWarning)
+                loop = asyncio.get_event_loop()
+        except RuntimeError:
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
+        return loop.run_until_complete(co)
+
+    @staticmethod
+    def _inject_type_overrides(
+        description: dict[str, Any],
+        schema_overrides: SchemaDict,
+    ) -> SchemaDict:
+        """
+        Attempt basic dtype inference from a cursor description.
+
+        Notes
+        -----
+        This is limited; the `type_code` property may contain almost anything,
+        from strings or python types to driver-specific codes, classes, enums, etc.
+        We currently only do the additional inference from string/python type values.
+        (Further refinement will require per-driver module knowledge and lookups).
+        """
+        dtype: PolarsDataType | None = None
+        for nm, desc in description.items():
+            if desc is None:
+                continue
+            elif nm not in schema_overrides:
+                type_code, _disp_size, internal_size, prec, scale, _null_ok = desc
+                if isclass(type_code):
+                    # python types, eg: int, float, str, etc
+                    with suppress(TypeError):
+                        dtype = _map_py_type_to_dtype(type_code)  # type: ignore[arg-type]
+
+                elif isinstance(type_code, str):
+                    # database/sql type names, eg: "VARCHAR", "NUMERIC", "BLOB", etc
+                    dtype = _infer_dtype_from_database_typename(
+                        value=type_code,
+                        raise_unmatched=False,
+                    )
+
+                if dtype is not None:
+                    # check additional cursor attrs to improve dtype specification
+                    if dtype == Float64 and internal_size == 4:
+                        dtype = Float32
+
+                    elif dtype in INTEGER_DTYPES and internal_size in (2, 4, 8):
+                        bits = internal_size * 8
+                        dtype = _integer_dtype_from_nbits(
+                            bits,
+                            unsigned=(dtype in UNSIGNED_INTEGER_DTYPES),
+                            default=dtype,
+                        )
+                    elif (
+                        dtype == Decimal
+                        and isinstance(prec, int)
+                        and isinstance(scale, int)
+                        and prec <= 38
+                        and scale <= 38
+                    ):
+                        dtype = Decimal(prec, scale)
+
+                if dtype is not None:
+                    schema_overrides[nm] = dtype  # type: ignore[index]
+
+        return schema_overrides
+
+    def _normalise_cursor(self, conn: Any) -> Cursor:
+        """Normalise a connection object such that we have the query executor."""
+        if self.driver_name == "sqlalchemy":
+            conn_type = type(conn).__name__
+            self.can_close_cursor = conn_type.endswith("Engine")
+
+            if conn_type in ("Session", "async_sessionmaker"):
+                return conn
+            else:
+                # where possible, use the raw connection to access arrow integration
+                if conn.engine.driver == "databricks-sql-python":
+                    self.driver_name = "databricks"
+                    return conn.engine.raw_connection().cursor()
+                elif conn.engine.driver == "duckdb_engine":
+                    self.driver_name = "duckdb"
+                    return conn.engine.raw_connection().driver_connection.c
+                elif conn_type in ("AsyncEngine", "Engine"):
+                    return conn.connect()
+                else:
+                    return conn
+
+        elif hasattr(conn, "cursor"):
+            # connection has a dedicated cursor; prefer over direct execute
+            cursor = cursor() if callable(cursor := conn.cursor) else cursor
+            self.can_close_cursor = True
+            return cursor
+
+        elif hasattr(conn, "execute"):
+            # can execute directly (given cursor, sqlalchemy connection, etc)
+            return conn
+
+        msg = f"Unrecognised connection {conn!r}; unable to find 'execute' method"
+        raise TypeError(msg)
+
+    async def _sqlalchemy_async_execute(self, query: TextClause, **options: Any) -> Any:
+        """Execute a query using an async SQLAlchemy connection."""
+        is_session = type(self.cursor).__name__ == "async_sessionmaker"
+        cursor = self.cursor.begin() if is_session else self.cursor  # type: ignore[attr-defined]
+        async with cursor as conn:
+            result = await conn.execute(query, **options)
+            return result
+
+    def _sqlalchemy_setup(
+        self, query: str | TextClause | Selectable, options: dict[str, Any]
+    ) -> tuple[Any, dict[str, Any], str | TextClause | Selectable]:
+        """Prepare a query for execution using a SQLAlchemy connection."""
+        from sqlalchemy.orm import Session
+        from sqlalchemy.sql import text
+        from sqlalchemy.sql.elements import TextClause
+
+        is_async = type(self.cursor).__name__ in (
+            "AsyncConnection",
+            "async_sessionmaker",
+        )
+        param_key = "parameters"
+        cursor_execute = None
+        if (
+            isinstance(self.cursor, Session)
+            and "parameters" in options
+            and "params" not in options
+        ):
+            options = options.copy()
+            options["params"] = options.pop("parameters")
+            param_key = "params"
+
+        params = options.get(param_key)
+        if (
+            not is_async
+            and isinstance(params, Sequence)
+            and hasattr(self.cursor, "exec_driver_sql")
+        ):
+            cursor_execute = self.cursor.exec_driver_sql
+            if isinstance(query, TextClause):
+                query = str(query)
+            if isinstance(params, list) and not all(
+                isinstance(p, (dict, tuple)) for p in params
+            ):
+                options[param_key] = tuple(params)
+
+        elif isinstance(query, str):
+            query = text(query)
+
+        if cursor_execute is None:
+            cursor_execute = (
+                self._sqlalchemy_async_execute if is_async else self.cursor.execute
+            )
+        return cursor_execute, options, query
+
+    def execute(
+        self,
+        query: str | TextClause | Selectable,
+        *,
+        options: dict[str, Any] | None = None,
+        select_queries_only: bool = True,
+    ) -> Self:
+        """Execute a query and reference the result set."""
+        if select_queries_only and isinstance(query, str):
+            q = re.search(r"\w{3,}", re.sub(r"/\*(.|[\r\n])*?\*/", "", query))
+            if (query_type := "" if not q else q.group(0)) in _INVALID_QUERY_TYPES:
+                msg = f"{query_type} statements are not valid 'read' queries"
+                raise UnsuitableSQLError(msg)
+
+        options = options or {}
+
+        if self.driver_name == "sqlalchemy":
+            cursor_execute, options, query = self._sqlalchemy_setup(query, options)
+        else:
+            cursor_execute = self.cursor.execute
+
+        # note: some cursor execute methods (eg: sqlite3) only take positional
+        # params, hence the slightly convoluted resolution of the 'options' dict
+        try:
+            params = signature(cursor_execute).parameters
+        except ValueError:
+            params = {}  # type: ignore[assignment]
+
+        if not options or any(
+            p.kind in (Parameter.KEYWORD_ONLY, Parameter.POSITIONAL_OR_KEYWORD)
+            for p in params.values()
+        ):
+            result = cursor_execute(query, **options)
+        else:
+            positional_options = (
+                options[o] for o in (params or options) if (not options or o in options)
+            )
+            result = cursor_execute(query, *positional_options)
+
+        # note: some cursors execute in-place
+        result = self.cursor if result is None else result
+        self.result = result
+        return self
+
+    def to_polars(
+        self,
+        *,
+        iter_batches: bool = False,
+        batch_size: int | None = None,
+        schema_overrides: SchemaDict | None = None,
+        infer_schema_length: int | None = N_INFER_DEFAULT,
+    ) -> DataFrame | Iterable[DataFrame]:
+        """
+        Convert the result set to a DataFrame.
+
+        Wherever possible we try to return arrow-native data directly; only
+        fall back to initialising with row-level data if no other option.
+        """
+        if self.result is None:
+            msg = "Cannot return a frame before executing a query"
+            raise RuntimeError(msg)
+        elif iter_batches and not batch_size:
+            msg = (
+                "Cannot set `iter_batches` without also setting a non-zero `batch_size`"
+            )
+            raise ValueError(msg)
+
+        for frame_init in (
+            self._from_arrow,  # init from arrow-native data (where support exists)
+            self._from_rows,  # row-wise fallback (sqlalchemy, dbapi2, pyodbc, etc)
+        ):
+            frame = frame_init(
+                batch_size=batch_size,
+                iter_batches=iter_batches,
+                schema_overrides=schema_overrides,
+                infer_schema_length=infer_schema_length,
+            )
+            if frame is not None:
+                return frame
+
+        msg = (
+            f"Currently no support for {self.driver_name!r} connection {self.cursor!r}"
+        )
+        raise NotImplementedError(msg)
