+++ py-polars/polars/io/spreadsheet/__init__.py
@@ -1,3 +1,6 @@
 from polars.io.spreadsheet.functions import read_excel, read_ods
 
-__all__ = ["read_excel", "read_ods"]
+__all__ = [
+    "read_excel",
+    "read_ods",
+]
